+++ tests/database/postgres/test_query_factory.py
@@ -168,9 +168,9 @@ def test_get_update_table_fake_column(column_strategy_list):
     update_table_all = query_factory.get_update_table("seed_table", UpdateColumnsTableStrategy("anon_table", column_strategy_list))
 
     assert update_table_all == [
-            "UPDATE anon_table SET "
-            "test_column1 = ( SELECT first_name FROM seed_table ORDER BY RANDOM() LIMIT 1),"
-            "test_column2 = ( SELECT last_name FROM seed_table ORDER BY RANDOM() LIMIT 1),"
+            "UPDATE anon_table AS \"updatetarget\" SET "
+            "test_column1 = ( SELECT first_name FROM seed_table WHERE \"updatetarget\"=\"updatetarget\" ORDER BY RANDOM() LIMIT 1),"
+            "test_column2 = ( SELECT last_name FROM seed_table WHERE \"updatetarget\"=\"updatetarget\" ORDER BY RANDOM() LIMIT 1),"
             "test_column3 = (''),"
             "test_column4 = ( SELECT md5(random()::text) ),"
             "test_column5 = ( SELECT CONCAT(md5(random()::text), '@', md5(random()::text), '.com') ),"
@@ -185,5 +185,5 @@ def test_get_update_table_literal(literal_strategy):
     ]))
 
     assert result_queries == [
-        "UPDATE anon_table SET literal_column = RANDOM();"
+        "UPDATE anon_table AS \"updatetarget\" SET literal_column = RANDOM();"
     ]